Fix std::isdigit calls to pick the right overload
There are two overloaded std::isidigit calls in the C/C++ stdlib.
The older C single int parameter one from <cctypes> and the new C++
template version which takes the character and a locale. Cast to the
char16_t character to an int to force the use of the libc version.
This assumes that we only care about digits 0-9 and aren't supporting
other digit sets.